refactor(kernel): Move TxtNode interface to @textlint/ast-node-types
