+++ packages/@textlint/ast-node-types/src/TextLintASTNodeTypes.ts
@@ -27,3 +27,41 @@ export const ASTNodeTypes = {
     Code: "Code",
     Delete: "Delete"
 };
+
+// TextLint AST Node
+export interface TxtNode {
+    type: keyof typeof ASTNodeTypes | string;
+    raw: string;
+    range: [number, number];
+    loc: TxtNodeLineLocation;
+    // parent is runtime information
+    // Not need in AST
+    parent?: TxtNode;
+}
+
+// Inline Node
+export interface TxtTextNode extends TxtNode {
+    value: string;
+}
+
+// Parent Node
+export interface TxtParentNode extends TxtNode {
+    children: TxtNode[] | TxtTextNode[];
+}
+
+export interface TxtRootNode extends TxtNode {
+    type: "Document";
+    children: TxtNode[];
+}
+
+export interface TxtNodeLineLocation {
+    start: TxtNodePosition;
+    end: TxtNodePosition;
+}
+
+export interface TxtNodePosition {
+    line: number; // start with 1
+    column: number; // start with 0
+    // This is for compatibility with JavaScript AST.
+    // https://gist.github.com/azu/8866b2cb9b7a933e01fe
+}
