+++ packages/@textlint/kernel/src/core/source-location.ts
@@ -2,7 +2,8 @@
 "use strict";
 import SourceCode from "./source-code";
 import RuleError, { RuleErrorPadding } from "./rule-error";
-import { TextlintMessage, TxtNode } from "../textlint-kernel-interface";
+import { TextlintMessage } from "../textlint-kernel-interface";
+import { TxtNode } from "@textlint/ast-node-types";
 
 const assert = require("assert");
 const ObjectAssign = require("object-assign");
