+++ packages/@textlint/kernel/src/task/textlint-core-task.ts
@@ -12,7 +12,8 @@ import MessageType from "../shared/type/MessageType";
 import { EventEmitter } from "events";
 import * as assert from "assert";
 import SourceCode from "../core/source-code";
-import { TextLintFixCommand, TextlintRuleOptions, TxtNode } from "../textlint-kernel-interface";
+import { TxtNode } from "@textlint/ast-node-types";
+import { TextLintFixCommand, TextlintRuleOptions } from "../textlint-kernel-interface";
 import { default as RuleContext, RuleReportedObject } from "../core/rule-context";
 import { RuleCreatorReporter } from "../core/rule-creator-helper";
 import FilterRuleContext from "../core/filter-rule-context";
