+++ packages/@textlint/kernel/src/textlint-kernel-interface.ts
@@ -1,5 +1,5 @@
 // rule config
-import { ASTNodeTypes } from "@textlint/ast-node-types";
+import { TxtNode } from "@textlint/ast-node-types";
 import { SeverityLevelTypes } from "./shared/type/SeverityLevel";
 import { TextLintRuleCreator } from "./core/rule-creator-helper";
 
@@ -47,44 +47,6 @@ export interface TextlintConfig {
      * quite options
      */
     quiet?: boolean;
-}
-
-// TextLint AST Node
-export interface TxtNode {
-    type: keyof typeof ASTNodeTypes | string;
-    raw: string;
-    range: [number, number];
-    loc: LineLocation;
-    // parent is runtime information
-    // Not need in AST
-    parent?: TxtNode;
-}
-
-// Inline Node
-export interface TxtTextNode extends TxtNode {
-    value: string;
-}
-
-// Parent Node
-export interface TxtParentNode extends TxtNode {
-    children: TxtNode[] | TxtTextNode[];
-}
-
-export interface TxtRootNode extends TxtNode {
-    type: "Document";
-    children: TxtNode[];
-}
-
-export interface LineLocation {
-    start: Position;
-    end: Position;
-}
-
-export interface Position {
-    line: number; // start with 1
-    column: number; // start with 0
-    // This is for compatibility with JavaScript AST.
-    // https://gist.github.com/azu/8866b2cb9b7a933e01fe
 }
 
 // Plugin
